remove replace default exports from indexes with the actual component
